Restructure code slightly for token scanning
